finished up twitter DB stuff, added javadocs

+++ app/src/main/java/org/sigma/photostream/data/Identifiable.java
@@ -4,5 +4,5 @@ package org.sigma.photostream.data;
  * Created by mattress on 2/18/2016.
  */
 public interface Identifiable {
-    Integer getID();
+    long getID();
 }
